refactor(types): move output-integration hook contracts to src/types
UseOutputChannelsReturn and UseTaskOutputIntegrationReturn now live in
types/workstation/output.ts; useGitOutputIntegration/types.ts moves to
types/workstation/gitOutputIntegration.ts. store/ atoms and services/guiAgent
no longer import from src/hooks (inverted edges), and the hooks import their
own return contracts from types/.
