+++ src/hooks/workStation/output/index.ts
@@ -1,6 +1,4 @@
 export { useOutputChannels } from "./useOutputChannels";
-export type { UseOutputChannelsReturn } from "./useOutputChannels";
 
 export { useFileWatchOutputIntegration } from "./useFileWatchOutputIntegration";
 export { useTaskOutputIntegration } from "./useTaskOutputIntegration";
-export type { UseTaskOutputIntegrationReturn } from "./useTaskOutputIntegration";
